[CAZB-3001] Update chargeable-vehicles endpoint to use cachedCharge

+++ src/main/java/uk/gov/caz/psr/dto/ChargeableAccountVehiclesResult.java
@@ -3,6 +3,7 @@ package uk.gov.caz.psr.dto;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonFormat.Shape;
 import io.swagger.annotations.ApiModelProperty;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.Comparator;
 import java.util.List;
@@ -17,7 +18,7 @@ import uk.gov.caz.psr.model.EntrantPayment;
 import uk.gov.caz.psr.repository.exception.CleanAirZoneNotFoundException;
 
 /**
- * Class that represents the returned JSON when client asks for account vehicles 
+ * Class that represents the returned JSON when client asks for account vehicles
  * that are chargeable.
  */
 @Value
@@ -27,7 +28,7 @@ public class ChargeableAccountVehiclesResult {
   /**
    * A list of vrns with associated charge, tariff code and paid travel dates.
    */
-  @ApiModelProperty(value = 
+  @ApiModelProperty(value =
       "${swagger.model.descriptions.chargeable-account-vehicles-result.results}")
   List<VrnWithTariffAndEntrancesPaid> results;
 
@@ -35,7 +36,7 @@ public class ChargeableAccountVehiclesResult {
    * Converts provided collection to {@link ChargeableAccountVehiclesResult}.
    * @param vrnsWithTariffAndCharge list of {@link VrnWithTariffAndEntrancesPaid}
    */
-  public static ChargeableAccountVehiclesResult from(Map<String, List<EntrantPayment>> results, 
+  public static ChargeableAccountVehiclesResult from(Map<String, List<EntrantPayment>> results,
       List<VrnWithTariffAndEntrancesPaid> vrnsWithTariffAndCharge) {
     List<VrnWithTariffAndEntrancesPaid> mappedResults = vrnsWithTariffAndCharge
         .stream()
@@ -46,7 +47,7 @@ public class ChargeableAccountVehiclesResult {
 
     return ChargeableAccountVehiclesResult.builder().results(mappedResults).build();
   }
-  
+
   /**
    * Builds {@link VrnWithTariffAndEntrancesPaid} based on compliance outcome.
    * @param complianceOutcome object containing vrn, charge and tariff
@@ -60,11 +61,11 @@ public class ChargeableAccountVehiclesResult {
         .orElseThrow(() -> new CleanAirZoneNotFoundException(cleanAirZoneId));
     return VrnWithTariffAndEntrancesPaid.builder()
         .vrn(complianceOutcome.getRegistrationNumber())
-        .charge(zoneOutcome.getCharge())
+        .charge(BigDecimal.valueOf(zoneOutcome.getCharge()))
         .tariffCode(zoneOutcome.getTariffCode())
         .build();
   }
-  
+
   /**
    * Enhances a given result with entrant payments.
    */
@@ -94,19 +95,19 @@ public class ChargeableAccountVehiclesResult {
      */
     @ApiModelProperty(value = "${swagger.model.descriptions.vrn-with-tariff-entrances.vrn}")
     String vrn;
-    
+
     /**
      * The tariff code for the vehicle in a given Clean Air Zone.
      */
-    @ApiModelProperty(value = 
+    @ApiModelProperty(value =
         "${swagger.model.descriptions.vrn-with-tariff-entrances.tariff-code}")
     String tariffCode;
-    
+
     /**
      * The charge incurred by the vehicle in a given Clean Air Zone.
      */
     @ApiModelProperty(value = "${swagger.model.descriptions.vrn-with-tariff-entrances.charge}")
-    double charge;
+    BigDecimal charge;
 
     /**
      * A list of days which are already paid.
@@ -115,5 +116,5 @@ public class ChargeableAccountVehiclesResult {
     @JsonFormat(shape = Shape.STRING, pattern = "yyyy-MM-dd")
     List<LocalDate> paidDates;
   }
-  
+
 }
